feat: 인증인가 엔드포인트 로그인, 사용자정보조회 시 company 정보도 같이 응답데이터로 반환
