+++ src/utils/HttpError.ts
@@ -1,0 +1,11 @@
+class HttpError extends Error {
+  public status: number;
+  constructor(message: string, status: number) {
+    super(message);
+    this.name = 'HttpError';
+    this.status = status;
+    Object.setPrototypeOf(this, HttpError.prototype);
+  }
+}
+
+export default HttpError;
